+++ src/utils/mod.rs
@@ -10,6 +10,7 @@ pub mod misc;
 pub mod monitoring;
 pub mod net;
 pub mod proofs_api;
+pub mod stream;
 pub mod version;
 
 use futures::{